api/config: Allow accessing ingest config without request
